29/11/2023 18:49

+++ src/main/java/com/piltover/repository/revenue/RevenueRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.piltover.dto.response.RevenueRes;
 
-public interface RevenueRepository extends JpaRepository<RevenueRes, Long>{
+public interface RevenueRepository extends JpaRepository<RevenueRes, Double> {
 	@Procedure("CallTotalRevenue")
 	List<RevenueRes> CallTotalRevenue(@Param("startDate") String startDate, @Param("endDate") String endDate);
 }
